refactor(router-plugin): follow unplugin standards and move lib code into "core"
