+++ packages/router-plugin/src/core/config.ts
