+++ packages/router-plugin/src/core/constants.ts
