+++ packages/router-plugin/src/core/router-generator-plugin.ts
