+++ packages/router-plugin/src/index.ts
@@ -1,4 +1,4 @@
-export { unpluginRouterCodeSplitterFactory } from './router-code-splitter'
-export { unpluginRouterGeneratorFactory } from './router-generator'
-export { configSchema } from './config'
-export type { Config } from './config'
+export { configSchema } from './core/config'
+export { unpluginRouterCodeSplitterFactory } from './core/router-code-splitter-plugin'
+export { unpluginRouterGeneratorFactory } from './core/router-generator-plugin'
+export type { Config } from './core/config'
